Moved class ids utils to separate module.
