+++ src/engine/core/managers/surge/utils/surge_generic.ts
@@ -1,5 +1,8 @@
 import { registry } from "@/engine/core/database";
+import { surgeConfig } from "@/engine/core/managers/surge/SurgeConfig";
+import { Squad } from "@/engine/core/objects/server/squad";
 import { LuaLogger } from "@/engine/core/utils/logging";
+import { TName } from "@/engine/lib/types";
 
 const logger: LuaLogger = new LuaLogger($filename);
 
@@ -15,4 +18,23 @@ export function launchSurgeSignalRockets(): void {
       signalLight.launch();
     }
   }
+}
+
+/**
+ * Check whether surge is enabled on the level.
+ *
+ * @returns whether surge can be started on provided level.
+ */
+export function isSurgeEnabledOnLevel(levelName: TName): boolean {
+  return surgeConfig.SURGE_DISABLED_LEVELS.get(levelName) !== true;
+}
+
+/**
+ * Check whether squad is immune to surge damage.
+ *
+ * @param object - squad object to check
+ * @returns whether provided community squad is immune to surge.
+ */
+export function isImmuneToSurgeObject(object: Squad): boolean {
+  return surgeConfig.IMMUNE_SQUAD_COMMUNITIES.get(object.faction) === true;
 }
