+++ src/engine/core/managers/surge/utils/surge_kill.ts
@@ -10,9 +10,9 @@ import {
   getNearestAvailableSurgeCover,
   getOnlineSurgeCoversList,
 } from "@/engine/core/managers/surge/utils/surge_cover";
+import { isImmuneToSurgeObject } from "@/engine/core/managers/surge/utils/surge_generic";
 import { pickSectionFromCondList } from "@/engine/core/utils/ini";
 import { LuaLogger } from "@/engine/core/utils/logging";
-import { isImmuneToSurgeObject } from "@/engine/core/utils/object/object_check";
 import { hasAlifeInfo } from "@/engine/core/utils/object/object_info_portion";
 import { isObjectOnLevel } from "@/engine/core/utils/object/object_location";
 import { animations, postProcessors } from "@/engine/lib/constants/animation";
