+++ src/engine/core/objects/ai/state/weapon/EvaluatorWeaponSet.ts
@@ -3,8 +3,8 @@ import { LuabindClass, property_evaluator } from "xray16";
 import { StalkerStateManager } from "@/engine/core/objects/ai/state/StalkerStateManager";
 import { states } from "@/engine/core/objects/animation/states";
 import { EWeaponAnimation } from "@/engine/core/objects/animation/types";
+import { isStrappableWeapon } from "@/engine/core/utils/class_ids";
 import { LuaLogger } from "@/engine/core/utils/logging";
-import { isStrappableWeapon } from "@/engine/core/utils/object";
 import { ClientObject, Optional } from "@/engine/lib/types";
 
 const logger: LuaLogger = new LuaLogger($filename);
