+++ src/engine/core/utils/job/job_create/job_create_object.ts
@@ -1,6 +1,6 @@
+import { isStalker } from "@/engine/core/utils/class_ids";
 import { IObjectJobDescriptor } from "@/engine/core/utils/job/job_types";
 import { LuaLogger } from "@/engine/core/utils/logging";
-import { isStalker } from "@/engine/core/utils/object/object_class";
 import { NIL } from "@/engine/lib/constants/words";
 import { ESchemeType, ServerCreatureObject } from "@/engine/lib/types";
