+++ src/engine/core/utils/object/index.ts
@@ -2,7 +2,6 @@ export * from "@/engine/core/utils/object/object_action";
 export * from "@/engine/core/utils/object/object_alife";
 export * from "@/engine/core/utils/object/object_anomaly";
 export * from "@/engine/core/utils/object/object_check";
-export * from "@/engine/core/utils/object/object_class";
 export * from "@/engine/core/utils/object/object_danger";
 export * from "@/engine/core/utils/object/object_find";
 export * from "@/engine/core/utils/object/object_get";
