+++ src/engine/lib/constants/class_ids.ts
@@ -1,6 +1,6 @@
 /* eslint sort-keys-fix/sort-keys-fix: "error" */
 
-import { createClassIds } from "@/engine/core/utils/class_ids";
+import { createClassIds } from "@/engine/core/utils/class_ids_list";
 import { TClassId } from "@/engine/lib/types";
 
 /**
