+++ src/engine/scripts/declarations/effects/world.ts
@@ -10,8 +10,9 @@ import { AnomalyZoneBinder } from "@/engine/core/objects/binders";
 import { SmartTerrain } from "@/engine/core/objects/server/smart_terrain";
 import { abort } from "@/engine/core/utils/assertion";
 import { extern } from "@/engine/core/utils/binding";
+import { isStalker } from "@/engine/core/utils/class_ids";
 import { LuaLogger } from "@/engine/core/utils/logging";
-import { isStalker, spawnItemsForObject } from "@/engine/core/utils/object";
+import { spawnItemsForObject } from "@/engine/core/utils/object";
 import { createVector } from "@/engine/core/utils/vector";
 import { TCommunity } from "@/engine/lib/constants/communities";
 import { questItems } from "@/engine/lib/constants/items/quest_items";
